refactor: migrate Money from double to BigDecimal
Money used a double as a deliberate trade-off to focus early on the hexagonal architecture. This adresses that debt following review feedback. A banking kata should ship BigDecimal from the start, not retrofit it.

Money is now BigDecimal-backed, scale normalized to 2 decimals, rejecting negative amounts and >2 decimal places. Domain, persistence and REST layers are updated accordingly, with extended coverage.

+++ infrastructure/src/main/java/com/bankingkata/adapter/in/rest/request/CreateAccountRequest.java
@@ -1,8 +1,10 @@
 package com.bankingkata.adapter.in.rest.request;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
 public class CreateAccountRequest {
-    private double initialBalance;
+    private BigDecimal initialBalance;
 }
